requirements.txt
fastapi>=0.115.0,<1.0
uvicorn>=0.34.0,<1.0
pydantic>=2.0,<3.0
pydantic-settings>=2.0,<3.0
requests>=2.31.0,<3.0
boto3>=1.35.0,<2.0
modal>=0.73.0
python-dotenv>=1.0,<2.0
pytest>=8.0,<9.0
